+++ workers/tests/integrationtests/test_api_indicator.py
@@ -2,9 +2,7 @@
 Testing FastAPI Applications:
 https://fastapi.tiangolo.com/tutorial/testing/
 """
-# API request tests are seperated for indicator and report
-# because of a bug when using two schemata.
-
+import os
 import unittest
 from typing import Optional
 
@@ -13,7 +11,7 @@ from schema import Schema
 
 from ohsome_quality_analyst.api.api import app
 
-from .api_response_schema import get_feature_schema, get_response_schema
+from .api_response_schema import get_general_schema, get_indicator_feature_schema
 from .utils import oqt_vcr
 
 
@@ -28,12 +26,12 @@ class TestApiIndicator(unittest.TestCase):
         self.feature_id = "3"
         self.fid_field = "ogc_fid"
 
-        self.response_schema = get_response_schema()
-        self.feature_schema = get_feature_schema()
+        self.general_schema = get_general_schema()
+        self.feature_schema = get_indicator_feature_schema()
 
     def run_tests(self, response) -> None:
         self.assertEqual(response.status_code, 200)
-        for schema in (self.response_schema, self.feature_schema):
+        for schema in (self.general_schema, self.feature_schema):
             self.validate(response.json(), schema)
 
     def validate(self, geojson: dict, schema: Schema) -> None:
@@ -135,6 +133,39 @@ class TestApiIndicator(unittest.TestCase):
             self.post_response(*parameters),
         ):
             self.run_tests(response)
+
+    @oqt_vcr.use_cassette()
+    def test_indicator_dataset_invalid(self):
+        data = {
+            "dataset": "foo",
+            "featureId": "3",
+        }
+        url = "/indicator/{0}".format(self.indicator_name)
+        response = self.client.post(url, json=data)
+        self.assertEqual(response.status_code, 422)
+
+    @oqt_vcr.use_cassette()
+    def test_indicator_invalid_set_of_arguments(self):
+
+        path = os.path.join(
+            os.path.dirname(os.path.abspath(__file__)),
+            "fixtures",
+            "heidelberg-altstadt-feature.geojson",
+        )
+        with open(path, "r") as f:
+            bpolys = f.read()
+        url = "/indicator/{0}".format(self.indicator_name)
+        for data in (
+            {
+                "bpolys": bpolys,
+                "dataset": "foo",
+                "featureId": "3",
+            },
+            {"dataset": "regions"},
+            {"feature_id": "3"},
+        ):
+            response = self.client.post(url, json=data)
+            self.assertEqual(response.status_code, 422)
 
 
 if __name__ == "__main__":