+++ workers/tests/integrationtests/test_api_indicator_geojson_io.py
@@ -2,8 +2,6 @@
 Testing FastAPI Applications:
 https://fastapi.tiangolo.com/tutorial/testing/
 """
-# API request tests are seperated for indicator and report
-# because of a bug when using two schemata.
 
 import os
 import unittest
@@ -15,9 +13,9 @@ from schema import Schema
 from ohsome_quality_analyst.api.api import app
 
 from .api_response_schema import (
-    get_feature_schema,
     get_featurecollection_schema,
-    get_response_schema,
+    get_general_schema,
+    get_indicator_feature_schema,
 )
 from .utils import oqt_vcr
 
@@ -35,8 +33,8 @@ class TestApiIndicatorIo(unittest.TestCase):
         self.indicator_name = "GhsPopComparisonBuildings"
         self.layer_name = "building_count"
 
-        self.response_schema = get_response_schema()
-        self.feature_schema = get_feature_schema()
+        self.general_schema = get_general_schema()
+        self.feature_schema = get_indicator_feature_schema()
         self.featurecollection_schema = get_featurecollection_schema()
 
     def run_tests(self, response, schemata: Tuple[Schema]) -> None:
@@ -67,13 +65,13 @@ class TestApiIndicatorIo(unittest.TestCase):
     def test_indicator_bpolys_geometry(self):
         geometry = get_fixture("heidelberg-altstadt-geometry.geojson")
         for response in (self.get_response(geometry), self.post_response(geometry)):
-            self.run_tests(response, (self.response_schema, self.feature_schema))
+            self.run_tests(response, (self.general_schema, self.feature_schema))
 
     @oqt_vcr.use_cassette()
     def test_indicator_bpolys_feature(self):
         feature = get_fixture("heidelberg-altstadt-feature.geojson")
         for response in (self.get_response(feature), self.post_response(feature)):
-            self.run_tests(response, (self.response_schema, self.feature_schema))
+            self.run_tests(response, (self.general_schema, self.feature_schema))
 
     @oqt_vcr.use_cassette()
     def test_indicator_bpolys_featurecollection(self):
@@ -85,7 +83,7 @@ class TestApiIndicatorIo(unittest.TestCase):
             self.post_response(featurecollection),
         ):
             self.run_tests(
-                response, (self.response_schema, self.featurecollection_schema)
+                response, (self.general_schema, self.featurecollection_schema)
             )
             for feature in response.json()["features"]:
                 self.validate(feature, self.feature_schema)
@@ -97,6 +95,20 @@ class TestApiIndicatorIo(unittest.TestCase):
             self.get_response(feature)
         with self.assertRaises(ValueError):
             self.post_response(feature)
+
+    @oqt_vcr.use_cassette()
+    def test_bpolys_invalid(self):
+        path = os.path.join(
+            os.path.dirname(os.path.abspath(__file__)),
+            "fixtures",
+            "invalid.geojson",
+        )
+        with open(path, "r") as file:
+            bpolys = file.read()
+        data = {"bpolys": bpolys, "layerName": self.layer_name}
+        url = f"/indicator/{self.indicator_name}"
+        response = self.client.post(url, json=data)
+        self.assertEqual(response.status_code, 422)
 
 
 if __name__ == "__main__":