+++ workers/tests/integrationtests/test_cli.py
@@ -5,7 +5,6 @@ https://click.palletsprojects.com/en/7.x/testing/?highlight=testing
 
 import unittest
 
-import geojson
 from click.testing import CliRunner
 
 from ohsome_quality_analyst.cli.cli import cli
@@ -125,7 +124,55 @@ class TestCliIntegration(unittest.TestCase):
             ["-q", "list-regions"],
         )
         self.assertEqual(result.exit_code, 0)
-        self.assertTrue(geojson.loads(result.output).is_valid)
+        self.assertTrue(result.output, str)
+
+    def test_get_indicator_layer_combination(self):
+        result = self.runner.invoke(
+            cli,
+            ["-q", "list-indicator-layer-combination"],
+        )
+        self.assertEqual(result.exit_code, 0)
+        self.assertIsInstance(result.output, str)
+
+    def test_list_indicators(self):
+        result = self.runner.invoke(
+            cli,
+            ["-q", "list-indicators"],
+        )
+        self.assertEqual(result.exit_code, 0)
+        self.assertIsInstance(result.output, str)
+
+    def test_list_reports(self):
+        result = self.runner.invoke(
+            cli,
+            ["-q", "list-reports"],
+        )
+        self.assertEqual(result.exit_code, 0)
+        self.assertIsInstance(result.output, str)
+
+    def test_list_layers(self):
+        result = self.runner.invoke(
+            cli,
+            ["-q", "list-layers"],
+        )
+        self.assertEqual(result.exit_code, 0)
+        self.assertIsInstance(result.output, str)
+
+    def test_list_datasets(self):
+        result = self.runner.invoke(
+            cli,
+            ["-q", "list-datasets"],
+        )
+        self.assertEqual(result.exit_code, 0)
+        self.assertIsInstance(result.output, str)
+
+    def test_list_fid_fields(self):
+        result = self.runner.invoke(
+            cli,
+            ["-q", "list-fid-fields"],
+        )
+        self.assertEqual(result.exit_code, 0)
+        self.assertIsInstance(result.output, str)
 
 
 if __name__ == "__main__":