+++ workers/tests/integrationtests/test_geodatabase.py
@@ -75,6 +75,12 @@ class TestGeodatabase(unittest.TestCase):
         self.assertIsNotNone(self.indicator.result.description)
         self.assertIsNotNone(self.indicator.result.svg)
 
+        # Test if data attributes were set
+        self.assertIsNotNone(self.indicator.pop_count)
+        self.assertIsNotNone(self.indicator.area)
+        self.assertIsNotNone(self.indicator.pop_count_per_sqkm)
+        self.assertIsNotNone(self.indicator.feature_count)
+
     def test_get_feature_ids(self):
         results = asyncio.run(db_client.get_feature_ids(self.dataset))
         self.assertIsInstance(results, list)
@@ -98,9 +104,13 @@ class TestGeodatabase(unittest.TestCase):
         with self.assertRaises(ValueError):
             asyncio.run(db_client.get_feature_from_db(self.dataset, "foo"))
 
-    def test_get_available_regions(self):
-        regions = asyncio.run(db_client.get_available_regions())
+    def test_get_regions_as_geojson(self):
+        regions = asyncio.run(db_client.get_regions_as_geojson())
         self.assertTrue(regions.is_valid)
+
+    def test_get_regions(self):
+        regions = asyncio.run(db_client.get_regions())
+        self.assertIsInstance(regions, list)
 
     def test_sanity_check_dataset(self):
         self.assertFalse(db_client.sanity_check_dataset("foo"))