+++ CHANGELOG.md
@@ -16,16 +16,27 @@
 
 - Added documentation on “How to add a layer definitions” ([#141])
 - Use concurrency when creating all indicators ([#153])
+- Add and improve tests for API ([#168])
+    - Add test cases covering the content of GeoJSON properties.
+    - Improve API response schemata by using less logic to create the schemata.
+- Add list indicator/layer combinations for API and CLI([#99])
+- Load indicator from DB will also load its data attributes ([#179])
+- Add parameter to 'regions' endpoint to only respond with a GeoJSON if requested ([#171])
+- Add csv output to list-region cli ([#65])
 
 ### How to upgrade?
 
 - If you set up your own database you will need to rebuild the database or delete the results table (`DROP TABLE results;`).
-
+[#65]: https://github.com/GIScience/ohsome-quality-analyst/issues/65
+[#99]: https://github.com/GIScience/ohsome-quality-analyst/issues/99
 [#102]: https://github.com/GIScience/ohsome-quality-analyst/pull/102
 [#106]: https://github.com/GIScience/ohsome-quality-analyst/issues/106
 [#141]: https://github.com/GIScience/ohsome-quality-analyst/pull/141
 [#149]: https://github.com/GIScience/ohsome-quality-analyst/pull/149
 [#153]: https://github.com/GIScience/ohsome-quality-analyst/pull/153
+[#168]: https://github.com/GIScience/ohsome-quality-analyst/pull/168
+[#171]: https://github.com/GIScience/ohsome-quality-analyst/issues/171
+[#179]: https://github.com/GIScience/ohsome-quality-analyst/pull/179
 [`pydantic`]: https://pydantic-docs.helpmanual.io/